+++ packages/working_router_lint/lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:analysis_server_plugin/plugin.dart';
 import 'package:analysis_server_plugin/registry.dart';
 
+import 'src/assists/remove_location_tree_element.dart';
 import 'src/assists/wrap_with_group.dart';
 import 'src/assists/wrap_with_shell.dart';
 
@@ -12,6 +13,7 @@ class _WorkingRouterPlugin extends Plugin {
 
   @override
   void register(PluginRegistry registry) {
+    registry.registerAssist(RemoveLocationTreeElement.new);
     registry.registerAssist(WrapWithGroup.new);
     registry.registerAssist(WrapWithShell.new);
   }